Refactor MOVE server into separate module, add controllers to coverage
in preparation for writing REST API tests.  This allows us to use `server.inject()` to test endpoints without going through the entire HTTP stack.
